+++ actions/src/encode.ts
@@ -1,93 +1,23 @@
-import {
-  toHex,
-  encodeAbiParameters,
-  parseAbiParameters,
-  Hex,
-  concatHex,
-} from "viem";
-import { FlowEdge, FlowMatrix, Stream } from "./circles";
+import { encodeAbiParameters, Hex, parseAbiParameters } from "viem";
+import { FlowMatrix } from "./circles/flowMatrix";
 
-const flowAbiType = {
-  name: "flow",
-  type: "tuple[]",
-  components: [
-    { name: "streamSinkId", type: "uint16" },
-    { name: "amount", type: "uint192" },
-  ],
-} as const;
+export const FlowMatrixAbi = parseAbiParameters(
+  "address[] flowVertices, (uint16 streamSinkId, uint192 amount)[] flowEdges, (uint16 sourceCoordinate, uint16[] flowEdgeIds, bytes data)[] streams, bytes packedCoordinates, uint16 sourceCoordinate",
+);
 
-const encodeFlowEdges = (flowEdges: FlowEdge[]) =>
-  encodeAbiParameters(
-    [flowAbiType],
-    [
-      flowEdges.map((edge) => ({
-        streamSinkId: edge.streamSinkId,
-        amount: BigInt(edge.amount),
-      })),
-    ],
-  );
-
-const streamAbiType = [
-  {
-    name: "streams",
-    type: "tuple[]",
-    components: [
-      { internalType: "uint16", name: "sourceCoordinate", type: "uint16" },
-      { internalType: "uint16[]", name: "flowEdgeIds", type: "uint16[]" },
-      { internalType: "bytes", name: "data", type: "bytes" },
-    ],
-  },
-] as const;
-
-const encodeStreams = (streams: Stream[]) =>
-  encodeAbiParameters(streamAbiType, [
-    streams.map((stream) => ({
-      ...stream,
-      data: toHex(stream.data), // Convert Uint8Array to hex string
-    })),
-  ]);
-
-export function encodeCallData(flowMatrix: FlowMatrix): Hex {
-  const {
+export function encodeCallData({
+  flowVertices,
+  flowEdges,
+  streams,
+  packedCoordinates,
+  sourceCoordinate,
+}: FlowMatrix): Hex {
+  return encodeAbiParameters(FlowMatrixAbi, [
+    // @ts-ignore
     flowVertices,
     flowEdges,
     streams,
     packedCoordinates,
     sourceCoordinate,
-  } = flowMatrix;
-  // This fucking shit doesn't work.
-  // const abiParams = [
-  //   parseAbiParameters("address[] flowVertices"),
-  //   flowAbiType,
-  //   streamAbiType,
-  //   parseAbiParameters("bytes packedCoordinates"),
-  //   parseAbiParameters("uint256 sourceCoordinate"),
-  // ];
-  // return encodeAbiParameters(abiParams, [
-  //   flowVertices,
-  //   flowEdges.map((edge) => ({
-  //     streamSinkId: edge.streamSinkId,
-  //     amount: BigInt(edge.amount),
-  //   })),
-  //   streams.map((stream) => ({
-  //     ...stream,
-  //     data: toHex(stream.data),
-  //   })),
-  //   packedCoordinates,
-  //   sourceCoordinate,
-  // ]);
-
-  return concatHex([
-    encodeAbiParameters(parseAbiParameters("address[] flowVertices"), [
-      flowVertices,
-    ]),
-    encodeFlowEdges(flowEdges),
-    encodeStreams(streams),
-    encodeAbiParameters(parseAbiParameters("bytes packedCoordinates"), [
-      packedCoordinates,
-    ]),
-    encodeAbiParameters(parseAbiParameters("uint256 sourceCoordinate"), [
-      BigInt(sourceCoordinate),
-    ]),
   ]);
 }
